use f64 instead of f32.

+++ src/camera.rs
@@ -10,10 +10,10 @@ pub struct Camera {
 }
 
 impl Camera {
-    pub fn new(ratio: f32) -> Camera {
-        let viewport_height = 2.0f32;
+    pub fn new(ratio: f64) -> Camera {
+        let viewport_height = 2.0f64;
         let viewport_width = ratio * viewport_height;
-        let focal_length = 1.0f32;
+        let focal_length = 1.0f64;
 
         let origin = Point3::zero();
         let horizontal = Vec3::new(viewport_width, 0.0, 0.0);
@@ -29,7 +29,7 @@ impl Camera {
         }
     }
 
-    pub fn get_ray(&self, u: f32, v: f32) -> Ray {
+    pub fn get_ray(&self, u: f64, v: f64) -> Ray {
         Ray::new(
             self.origin,
             self.lower_left_corner + u * self.horizontal + v * self.vertical - self.origin,
